restructured tests into test folder
